refactor: remove props
