+++ components/RightBarContainer/index.tsx
@@ -1,27 +1,10 @@
-import React, { useCallback } from "react";
+import React from "react";
 import { useAppSelector } from "@/lib/hooks";
-import { useChatContext } from "@/context/ChatContext";
 
 import RightBar from "../RightBar";
 // import GenAiImage from "./GenAiImage";
 
 export const RightBarContainer = () => {
-  const { keyword, setKeyword } = useChatContext();
-
-  // TODO: request dalle
-  const handleSubmit = useCallback(
-    (e: React.FormEvent<HTMLElement>) => {
-      e.preventDefault();
-      if (!keyword.trim()) {
-        setKeyword("");
-        return;
-      }
-      // TODO: to get image
-      setKeyword("");
-    },
-    [keyword, setKeyword]
-  );
-
   const isReceiving = useAppSelector((state) => state.genAi.isReceiving);
   const generatedImgUrls = useAppSelector(
     (state) => state.genAi.generatedImgUrls
@@ -29,7 +12,7 @@ export const RightBarContainer = () => {
 
   return (
     <>
-      <RightBar handleSubmit={handleSubmit} />
+      <RightBar />
 
       {/* <div
         style={{
